request: add context to http calls
