+++ request.go
@@ -9,11 +9,8 @@ import (
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
-	"log"
-	"net"
 	"net/http"
 	"net/url"
-	"reflect"
 	"sort"
 	"strconv"
 	"strings"
@@ -27,35 +24,25 @@ type Command interface {
 }
 
 // SyncCommand represents a CloudStack synchronous request
-type SyncCommand interface {
-	// CloudStack API command name
-	APIName() string
+type syncCommand interface {
+	Command
 	// Response interface to Unmarshal the JSON into
 	response() interface{}
 }
 
-// AsyncCommand represents a CloudStack asynchronous request
-type AsyncCommand interface {
-	// CloudStack API command name
-	APIName() string
+// asyncCommand represents a async CloudStack request
+type asyncCommand interface {
+	Command
 	// Response interface to Unmarshal the JSON into
 	asyncResponse() interface{}
 }
 
-// Command represents an action to be done on the params before sending them
+// onBeforeHook represents an action to be done on the params before sending them
 //
 // This little took helps with issue of relying on JSON serialization logic only.
 // `omitempty` may make sense in some cases but not all the time.
 type onBeforeHook interface {
 	onBeforeSend(params *url.Values) error
-}
-
-// AsyncInfo represents the details for any async call
-//
-// It retries at most Retries time and waits for Delay between each retry
-type AsyncInfo struct {
-	Retries int
-	Delay   int
 }
 
 const (
@@ -168,46 +155,18 @@ func (e *booleanSyncResponse) Error() error {
 	return fmt.Errorf("API error: %s", e.DisplayText)
 }
 
-type asyncJob struct {
-	command      AsyncCommand
-	delay        int
-	retries      int
-	responseChan chan<- *AsyncJobResult
+type syncJob struct {
+	command      syncCommand
+	responseChan chan<- interface{}
 	errorChan    chan<- error
 	ctx          context.Context
 }
 
-func csQuotePlus(s string) string {
-	s = strings.Replace(s, "+", "%20", -1)
-	s = strings.Replace(s, "%5B", "[", -1)
-	s = strings.Replace(s, "%5D", "]", -1)
-	return s
-}
-
-func csEncode(s string) string {
-	return csQuotePlus(url.QueryEscape(s))
-}
-
-func rawValue(b json.RawMessage) (json.RawMessage, error) {
-	var m map[string]json.RawMessage
-
-	if err := json.Unmarshal(b, &m); err != nil {
-		return nil, err
-	}
-	for _, v := range m {
-		return v, nil
-	}
-	return nil, nil
-}
-
-func rawValues(b json.RawMessage) (json.RawMessage, error) {
-	var i []json.RawMessage
-
-	if err := json.Unmarshal(b, &i); err != nil {
-		return nil, nil
-	}
-
-	return i[0], nil
+type asyncJob struct {
+	command      asyncCommand
+	responseChan chan<- *AsyncJobResult
+	errorChan    chan<- error
+	ctx          context.Context
 }
 
 func (exo *Client) parseResponse(resp *http.Response) (json.RawMessage, error) {
@@ -235,11 +194,35 @@ func (exo *Client) parseResponse(resp *http.Response) (json.RawMessage, error) {
 	return b, nil
 }
 
+func (exo *Client) processSyncJob(ctx context.Context, job *syncJob) {
+	defer close(job.responseChan)
+	defer close(job.errorChan)
+
+	body, err := exo.request(ctx, job.command.APIName(), job.command)
+	if err != nil {
+		job.errorChan <- err
+		return
+	}
+
+	resp := job.command.response()
+	if err := json.Unmarshal(body, resp); err != nil {
+		r := new(ErrorResponse)
+		if e := json.Unmarshal(body, r); e != nil {
+			job.errorChan <- r
+			return
+		}
+		job.errorChan <- err
+		return
+	}
+
+	job.responseChan <- resp.(interface{})
+}
+
 func (exo *Client) processAsyncJob(ctx context.Context, job *asyncJob) {
 	defer close(job.responseChan)
 	defer close(job.errorChan)
 
-	body, err := exo.request(job.command.APIName(), job.command)
+	body, err := exo.request(ctx, job.command.APIName(), job.command)
 	if err != nil {
 		job.errorChan <- err
 		return
@@ -262,18 +245,13 @@ func (exo *Client) processAsyncJob(ctx context.Context, job *asyncJob) {
 		return
 	}
 
-	for job.retries > 0 {
+	for iteration := 0; ; iteration++ {
 		select {
 		case <-ctx.Done():
 			job.errorChan <- ctx.Err()
 			return
 		default:
-			job.retries--
-
-			end, _ := ctx.Deadline()
-
-			when := end.Add(time.Duration(job.delay*-job.retries) * time.Second)
-			time.Sleep(when.Sub(time.Now()))
+			time.Sleep(exo.RetryStrategy(int64(iteration)))
 
 			req := &QueryAsyncJobResult{JobID: jobResult.JobID}
 			resp, err := exo.Request(req)
@@ -298,25 +276,15 @@ func (exo *Client) processAsyncJob(ctx context.Context, job *asyncJob) {
 			}
 		}
 	}
-
-	job.errorChan <- fmt.Errorf("Maximum number of retries reached")
 }
 
-// AsyncRequest performs an asynchronous request and polls it for retries * day [s]
-func (exo *Client) AsyncRequest(req AsyncCommand, async AsyncInfo) (interface{}, error) {
-	totalTime := time.Duration(async.Delay*async.Retries) * time.Second
-	end := time.Now().Add(totalTime)
-
-	ctx, cancel := context.WithDeadline(context.Background(), end)
-	defer cancel()
-
+// asyncRequest perform an asynchronous job with a context
+func (exo *Client) asyncRequest(ctx context.Context, req asyncCommand) (interface{}, error) {
 	responseChan := make(chan *AsyncJobResult, 1)
 	errorChan := make(chan error, 1)
 
 	go exo.processAsyncJob(ctx, &asyncJob{
 		command:      req,
-		delay:        async.Delay,
-		retries:      async.Retries,
 		responseChan: responseChan,
 		errorChan:    errorChan,
 		ctx:          ctx,
@@ -324,7 +292,6 @@ func (exo *Client) AsyncRequest(req AsyncCommand, async AsyncInfo) (interface{},
 
 	select {
 	case result := <-responseChan:
-		cancel()
 		resp := req.asyncResponse()
 		if err := json.Unmarshal(*(result.JobResult), resp); err != nil {
 			return nil, err
@@ -332,57 +299,84 @@ func (exo *Client) AsyncRequest(req AsyncCommand, async AsyncInfo) (interface{},
 		return resp, nil
 
 	case err := <-errorChan:
-		cancel()
 		return nil, err
 
 	case <-ctx.Done():
-		cancel()
 		err := <-errorChan
 		return nil, err
 	}
 }
 
-// BooleanRequest performs a sync request on a boolean call
-func (exo *Client) BooleanRequest(req SyncCommand) error {
+// syncRequest performs a sync request with a context
+func (exo *Client) syncRequest(ctx context.Context, req syncCommand) (interface{}, error) {
+	responseChan := make(chan interface{}, 1)
+	errorChan := make(chan error, 1)
+
+	go exo.processSyncJob(ctx, &syncJob{
+		command:      req,
+		responseChan: responseChan,
+		errorChan:    errorChan,
+		ctx:          ctx,
+	})
+
+	select {
+	case result := <-responseChan:
+		return result, nil
+	case err := <-errorChan:
+		return nil, err
+	case <-ctx.Done():
+		err := <-errorChan
+		return nil, err
+	}
+}
+
+// BooleanRequest performs the given boolean command
+func (exo *Client) BooleanRequest(req Command) error {
 	resp, err := exo.Request(req)
 	if err != nil {
 		return err
 	}
 
-	return resp.(*booleanSyncResponse).Error()
+	// CloudStack returns a different type between sync and async success responses
+	if b, ok := resp.(*booleanSyncResponse); ok {
+		return b.Error()
+	}
+	if b, ok := resp.(*booleanAsyncResponse); ok {
+		return b.Error()
+	}
+
+	panic(fmt.Errorf("The command %s is not a proper boolean response. %#v", req.APIName(), resp))
 }
 
-// BooleanAsyncRequest performs a sync request on a boolean call
-func (exo *Client) BooleanAsyncRequest(req AsyncCommand, async AsyncInfo) error {
-	resp, err := exo.AsyncRequest(req, async)
-	if err != nil {
-		return err
-	}
+// Request performs the given command
+func (exo *Client) Request(req Command) (interface{}, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), exo.Timeout)
+	defer cancel()
 
-	return resp.(*booleanAsyncResponse).Error()
+	switch req.(type) {
+	case syncCommand:
+		return exo.syncRequest(ctx, req.(syncCommand))
+	case asyncCommand:
+		return exo.asyncRequest(ctx, req.(asyncCommand))
+	default:
+		panic(fmt.Errorf("The command %s is not a proper Sync or Async command", req.APIName()))
+	}
 }
 
-// Request performs a sync request on a generic command
-func (exo *Client) Request(req SyncCommand) (interface{}, error) {
-	body, err := exo.request(req.APIName(), req)
-	if err != nil {
-		return nil, err
+// RequestWithContext preforms a request with a context
+func (exo *Client) RequestWithContext(ctx context.Context, req Command) (interface{}, error) {
+	switch req.(type) {
+	case syncCommand:
+		return exo.syncRequest(ctx, req.(syncCommand))
+	case asyncCommand:
+		return exo.asyncRequest(ctx, req.(asyncCommand))
+	default:
+		panic(fmt.Errorf("The command %s is not a proper Sync or Async command", req.APIName()))
 	}
-
-	resp := req.response()
-	if err := json.Unmarshal(body, resp); err != nil {
-		r := new(ErrorResponse)
-		if e := json.Unmarshal(body, r); e != nil {
-			return nil, r
-		}
-		return nil, err
-	}
-
-	return resp, nil
 }
 
 // request makes a Request while being close to the metal
-func (exo *Client) request(command string, req interface{}) (json.RawMessage, error) {
+func (exo *Client) request(ctx context.Context, command string, req interface{}) (json.RawMessage, error) {
 	params := url.Values{}
 	err := prepareValues("", &params, req)
 	if err != nil {
@@ -422,8 +416,16 @@ func (exo *Client) request(command string, req interface{}) (json.RawMessage, er
 	mac.Write([]byte(strings.ToLower(query)))
 	signature := csEncode(base64.StdEncoding.EncodeToString(mac.Sum(nil)))
 
-	reader := strings.NewReader(fmt.Sprintf("%s&signature=%s", csQuotePlus(query), signature))
-	resp, err := exo.client.Post(exo.endpoint, "application/x-www-form-urlencoded", reader)
+	payload := fmt.Sprintf("%s&signature=%s", csQuotePlus(query), signature)
+
+	request, err := http.NewRequest("POST", exo.endpoint, strings.NewReader(payload))
+	if err != nil {
+		return nil, err
+	}
+	request.Header.Add("Content-Type", "application/x-www-form-urlencoded")
+	request.Header.Add("Content-Length", strconv.Itoa(len(payload)))
+
+	resp, err := exo.client.Do(request.WithContext(ctx))
 	if err != nil {
 		return nil, err
 	}
@@ -435,211 +437,4 @@ func (exo *Client) request(command string, req interface{}) (json.RawMessage, er
 	}
 
 	return body, nil
-}
-
-// prepareValues uses a command to build a POST request
-//
-// command is not a Command so it's easier to Test
-func prepareValues(prefix string, params *url.Values, command interface{}) error {
-	value := reflect.ValueOf(command)
-	typeof := reflect.TypeOf(command)
-	// Going up the pointer chain to find the underlying struct
-	for typeof.Kind() == reflect.Ptr {
-		typeof = typeof.Elem()
-		value = value.Elem()
-	}
-
-	for i := 0; i < typeof.NumField(); i++ {
-		field := typeof.Field(i)
-		val := value.Field(i)
-		tag := field.Tag
-		if json, ok := tag.Lookup("json"); ok {
-			n, required := extractJSONTag(field.Name, json)
-			name := prefix + n
-
-			switch val.Kind() {
-			case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-				v := val.Int()
-				if v == 0 {
-					if required {
-						return fmt.Errorf("%s.%s (%v) is required, got 0", typeof.Name(), field.Name, val.Kind())
-					}
-				} else {
-					(*params).Set(name, strconv.FormatInt(v, 10))
-				}
-			case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-				v := val.Uint()
-				if v == 0 {
-					if required {
-						return fmt.Errorf("%s.%s (%v) is required, got 0", typeof.Name(), field.Name, val.Kind())
-					}
-				} else {
-					(*params).Set(name, strconv.FormatUint(v, 10))
-				}
-			case reflect.Float32, reflect.Float64:
-				v := val.Float()
-				if v == 0 {
-					if required {
-						return fmt.Errorf("%s.%s (%v) is required, got 0", typeof.Name(), field.Name, val.Kind())
-					}
-				} else {
-					(*params).Set(name, strconv.FormatFloat(v, 'f', -1, 64))
-				}
-			case reflect.String:
-				v := val.String()
-				if v == "" {
-					if required {
-						return fmt.Errorf("%s.%s (%v) is required, got \"\"", typeof.Name(), field.Name, val.Kind())
-					}
-				} else {
-					(*params).Set(name, v)
-				}
-			case reflect.Bool:
-				v := val.Bool()
-				if v == false {
-					if required {
-						params.Set(name, "false")
-					}
-				} else {
-					(*params).Set(name, "true")
-				}
-			case reflect.Ptr:
-				if val.IsNil() {
-					if required {
-						return fmt.Errorf("%s.%s (%v) is required, got empty ptr", typeof.Name(), field.Name, val.Kind())
-					}
-				} else {
-					switch field.Type.Elem().Kind() {
-					case reflect.Bool:
-						params.Set(name, strconv.FormatBool(val.Elem().Bool()))
-					default:
-						log.Printf("[SKIP] %s.%s (%v) not supported", typeof.Name(), field.Name, field.Type.Elem().Kind())
-					}
-				}
-			case reflect.Slice:
-				switch field.Type.Elem().Kind() {
-				case reflect.Uint8:
-					switch field.Type {
-					case reflect.TypeOf(net.IPv4zero):
-						ip := (net.IP)(val.Bytes())
-						if ip == nil || ip.Equal(net.IPv4zero) {
-							if required {
-								return fmt.Errorf("%s.%s (%v) is required, got zero IPv4 address", typeof.Name(), field.Name, val.Kind())
-							}
-						} else {
-							(*params).Set(name, ip.String())
-						}
-					default:
-						if val.Len() == 0 {
-							if required {
-								return fmt.Errorf("%s.%s (%v) is required, got empty slice", typeof.Name(), field.Name, val.Kind())
-							}
-						} else {
-							v := val.Bytes()
-							(*params).Set(name, base64.StdEncoding.EncodeToString(v))
-						}
-					}
-				case reflect.String:
-					{
-						if val.Len() == 0 {
-							if required {
-								return fmt.Errorf("%s.%s (%v) is required, got empty slice", typeof.Name(), field.Name, val.Kind())
-							}
-						} else {
-							elems := make([]string, 0, val.Len())
-							for i := 0; i < val.Len(); i++ {
-								// XXX what if the value contains a comma? Double encode?
-								s := val.Index(i).String()
-								elems = append(elems, s)
-							}
-							(*params).Set(name, strings.Join(elems, ","))
-						}
-					}
-				default:
-					if val.Len() == 0 {
-						if required {
-							return fmt.Errorf("%s.%s (%v) is required, got empty slice", typeof.Name(), field.Name, val.Kind())
-						}
-					} else {
-						err := prepareList(name, params, val.Interface())
-						if err != nil {
-							return err
-						}
-					}
-				}
-			case reflect.Map:
-				if val.Len() == 0 {
-					if required {
-						return fmt.Errorf("%s.%s (%v) is required, got empty map", typeof.Name(), field.Name, val.Kind())
-					}
-				} else {
-					err := prepareMap(name, params, val.Interface())
-					if err != nil {
-						return err
-					}
-				}
-			default:
-				if required {
-					return fmt.Errorf("Unsupported type %s.%s (%v)", typeof.Name(), field.Name, val.Kind())
-				}
-			}
-		} else {
-			log.Printf("[SKIP] %s.%s no json label found", typeof.Name(), field.Name)
-		}
-	}
-
-	return nil
-}
-
-func prepareList(prefix string, params *url.Values, slice interface{}) error {
-	value := reflect.ValueOf(slice)
-
-	for i := 0; i < value.Len(); i++ {
-		prepareValues(fmt.Sprintf("%s[%d].", prefix, i), params, value.Index(i).Interface())
-	}
-
-	return nil
-}
-
-func prepareMap(prefix string, params *url.Values, m interface{}) error {
-	value := reflect.ValueOf(m)
-
-	for i, key := range value.MapKeys() {
-		var keyName string
-		var keyValue string
-
-		switch key.Kind() {
-		case reflect.String:
-			keyName = key.String()
-		default:
-			return fmt.Errorf("Only map[string]string are supported (XXX)")
-		}
-
-		val := value.MapIndex(key)
-		switch val.Kind() {
-		case reflect.String:
-			keyValue = val.String()
-		default:
-			return fmt.Errorf("Only map[string]string are supported (XXX)")
-		}
-		params.Set(fmt.Sprintf("%s[%d].%s", prefix, i, keyName), keyValue)
-	}
-	return nil
-}
-
-// extractJSONTag returns the variable name or defaultName as well as if the field is required (!omitempty)
-func extractJSONTag(defaultName, jsonTag string) (string, bool) {
-	tags := strings.Split(jsonTag, ",")
-	name := tags[0]
-	required := true
-	for _, tag := range tags {
-		if tag == "omitempty" {
-			required = false
-		}
-	}
-
-	if name == "" || name == "omitempty" {
-		name = defaultName
-	}
-	return name, required
 }
