+++ request_test.go
@@ -4,284 +4,12 @@ import (
 	"encoding/json"
 	"fmt"
 	"log"
-	"net"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
 	"testing"
+	"time"
 )
-
-func TestPrepareValues(t *testing.T) {
-	type tag struct {
-		Name      string `json:"name"`
-		IsVisible bool   `json:"isvisible,omitempty"`
-	}
-
-	profile := struct {
-		IgnoreMe    string
-		Zone        string            `json:"myzone,omitempty"`
-		Name        string            `json:"name"`
-		NoName      string            `json:"omitempty"`
-		ID          int               `json:"id"`
-		UserID      uint              `json:"user_id"`
-		IsGreat     bool              `json:"is_great"`
-		IsAmazing   *bool             `json:"is_amazing,omitempty"`
-		Num         float64           `json:"num"`
-		Bytes       []byte            `json:"bytes"`
-		IDs         []string          `json:"ids,omitempty"`
-		TagPointers []*tag            `json:"tagpointers,omitempty"`
-		Tags        []tag             `json:"tags,omitempty"`
-		Map         map[string]string `json:"map"`
-		IP          net.IP            `json:"ip,omitempty"`
-	}{
-		IgnoreMe: "bar",
-		Name:     "world",
-		NoName:   "foo",
-		ID:       1,
-		UserID:   uint(2),
-		Num:      3.14,
-		Bytes:    []byte("exo"),
-		IDs:      []string{"1", "2", "three"},
-		TagPointers: []*tag{
-			{Name: "foo"},
-			{Name: "bar", IsVisible: false},
-		},
-		Tags: []tag{
-			{Name: "foo"},
-			{Name: "bar", IsVisible: false},
-		},
-		Map: map[string]string{
-			"foo": "bar",
-		},
-		IP: net.IPv4(192, 168, 0, 11),
-	}
-
-	params := url.Values{}
-	err := prepareValues("", &params, profile)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if is_amazing, ok := params["is_amazing"]; ok {
-		t.Errorf("is_amazing shouldn't be set, got %v", is_amazing)
-	}
-
-	if myzone, ok := params["myzone"]; ok {
-		t.Errorf("myzone params shouldn't be set, got %v", myzone)
-	}
-
-	if params.Get("NoName") != "foo" {
-		t.Errorf("NoName params wasn't properly set, got %v", params.Get("NoName"))
-	}
-
-	if params.Get("name") != "world" {
-		t.Errorf("name params wasn't properly set, got %v", params.Get("name"))
-	}
-
-	if params.Get("bytes") != "ZXhv" {
-		t.Errorf("bytes params wasn't properly encoded in base 64, got %v", params.Get("bytes"))
-	}
-
-	if params.Get("ids") != "1,2,three" {
-		t.Errorf("array of strings, wasn't property encoded, got %v", params.Get("ids"))
-	}
-
-	if _, ok := params["ignoreme"]; ok {
-		t.Errorf("IgnoreMe key was set")
-	}
-
-	v := params.Get("tags[0].name")
-	if v != "foo" {
-		t.Errorf("expected tags to be serialized as foo, got %#v", v)
-	}
-
-	v = params.Get("tagpointers[0].name")
-	if v != "foo" {
-		t.Errorf("expected tag pointers to be serialized as foo, got %#v", v)
-	}
-
-	v = params.Get("map[0].foo")
-	if v != "bar" {
-		t.Errorf("expected map to be serialized as .foo => \"bar\", got .foo => %#v", v)
-	}
-
-	v = params.Get("is_great")
-	if v != "false" {
-		t.Errorf("expected bool to be serialized as \"false\", got %#v", v)
-	}
-
-	v = params.Get("ip")
-	if v != "192.168.0.11" {
-		t.Errorf("expected ip to be serialized as \"192.168.0.11\", got %#v", v)
-	}
-}
-
-func TestPrepareValuesStringRequired(t *testing.T) {
-	profile := struct {
-		RequiredField string `json:"requiredfield"`
-	}{}
-
-	params := url.Values{}
-	err := prepareValues("", &params, &profile)
-	if err == nil {
-		t.Errorf("It should have failed")
-	}
-}
-
-func TestPrepareValuesBoolRequired(t *testing.T) {
-	profile := struct {
-		RequiredField bool `json:"requiredfield"`
-	}{}
-
-	params := url.Values{}
-	err := prepareValues("", &params, &profile)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if params.Get("requiredfield") != "false" {
-		t.Errorf("bool params wasn't set to false (default value)")
-	}
-}
-
-func TestPrepareValuesBoolPtrRequired(t *testing.T) {
-	profile := struct {
-		RequiredField *bool `json:"requiredfield"`
-	}{}
-
-	params := url.Values{}
-	err := prepareValues("", &params, &profile)
-
-	if err == nil {
-		t.Errorf("It should have failed")
-	}
-}
-
-func TestPrepareValuesIntRequired(t *testing.T) {
-	profile := struct {
-		RequiredField int64 `json:"requiredfield"`
-	}{}
-
-	params := url.Values{}
-	err := prepareValues("", &params, &profile)
-	if err == nil {
-		t.Errorf("It should have failed")
-	}
-}
-
-func TestPrepareValuesUintRequired(t *testing.T) {
-	profile := struct {
-		RequiredField uint64 `json:"requiredfield"`
-	}{}
-
-	params := url.Values{}
-	err := prepareValues("", &params, &profile)
-	if err == nil {
-		t.Errorf("It should have failed")
-	}
-}
-
-func TestPrepareValuesBytesRequired(t *testing.T) {
-	profile := struct {
-		RequiredField []byte `json:"requiredfield"`
-	}{}
-
-	params := url.Values{}
-	err := prepareValues("", &params, &profile)
-	if err == nil {
-		t.Errorf("It should have failed")
-	}
-}
-
-func TestPrepareValuesSliceString(t *testing.T) {
-	profile := struct {
-		RequiredField []string `json:"requiredfield"`
-	}{}
-
-	params := url.Values{}
-	err := prepareValues("", &params, &profile)
-	if err == nil {
-		t.Errorf("It should have failed")
-	}
-}
-
-func TestPrepareValuesIP(t *testing.T) {
-	profile := struct {
-		RequiredField net.IP `json:"requiredfield"`
-	}{}
-
-	params := url.Values{}
-	err := prepareValues("", &params, &profile)
-	if err == nil {
-		t.Errorf("It should have failed")
-	}
-}
-
-func TestPrepareValuesIPZero(t *testing.T) {
-	profile := struct {
-		RequiredField net.IP `json:"requiredfield"`
-	}{
-		RequiredField: net.IPv4zero,
-	}
-
-	params := url.Values{}
-	err := prepareValues("", &params, &profile)
-	if err == nil {
-		t.Errorf("It should have failed")
-	}
-}
-
-func TestPrepareValuesMap(t *testing.T) {
-	profile := struct {
-		RequiredField map[string]string `json:"requiredfield"`
-	}{}
-
-	params := url.Values{}
-	err := prepareValues("", &params, &profile)
-	if err == nil {
-		t.Errorf("It should have failed")
-	}
-}
-
-func TestPrepareValuesBoolPtr(t *testing.T) {
-	tru := new(bool)
-	f := new(bool)
-	*tru = true
-	*f = false
-
-	profile := struct {
-		IsOne   bool  `json:"is_one,omitempty"`
-		IsTwo   bool  `json:"is_two,omitempty"`
-		IsThree *bool `json:"is_three,omitempty"`
-		IsFour  *bool `json:"is_four,omitempty"`
-		IsFive  *bool `json:"is_five,omitempty"`
-	}{
-		IsOne:   true,
-		IsTwo:   false,
-		IsThree: tru,
-		IsFour:  f,
-	}
-
-	params := url.Values{}
-	err := prepareValues("", &params, &profile)
-	if err != nil {
-		t.Error(err)
-	}
-	if params["is_one"][0] != "true" {
-		t.Errorf("Expected is_one to be true")
-	}
-	if is_two, ok := params["is_two"]; ok {
-		t.Errorf("Expected is_two to be missing, got %v", is_two)
-	}
-	if params["is_three"][0] != "true" {
-		t.Errorf("Expected is_three to be true")
-	}
-	if params["is_four"][0] != "false" {
-		t.Errorf("Expected is_four to be false")
-	}
-	if is_five, ok := params["is_five"]; ok {
-		t.Errorf("Expected is_five to be missing, got %v", is_five)
-	}
-}
 
 func TestRequest(t *testing.T) {
 	params := url.Values{}
@@ -347,7 +75,7 @@ func TestBooleanAsyncRequest(t *testing.T) {
 	req := &ExpungeVirtualMachine{
 		ID: "123",
 	}
-	err := cs.BooleanAsyncRequest(req, AsyncInfo{Delay: 1, Retries: 1})
+	err := cs.BooleanRequest(req)
 
 	if err != nil {
 		t.Errorf(err.Error())
@@ -373,17 +101,17 @@ func TestBooleanAsyncRequestTimeout(t *testing.T) {
 	`)
 	defer ts.Close()
 
-	cs := NewClient(ts.URL, "TOKEN", "SECRET")
+	cs := NewClientWithTimeout(ts.URL, "TOKEN", "SECRET", time.Second)
 	req := &ExpungeVirtualMachine{
 		ID: "123",
 	}
-	err := cs.BooleanAsyncRequest(req, AsyncInfo{Delay: 1, Retries: 2})
+	err := cs.BooleanRequest(req)
 
 	if err == nil {
 		t.Error("An error was expected")
 	}
 
-	if err.Error() != "Maximum number of retries reached" {
+	if err.Error() != "context deadline exceeded" {
 		t.Errorf("Unexpected error message: %s", err.Error())
 	}
 }
