+++ types.go
@@ -2,6 +2,7 @@ package egoscale
 
 import (
 	"net/http"
+	"time"
 )
 
 // Client represents the CloudStack API client
@@ -10,7 +11,14 @@ type Client struct {
 	endpoint  string
 	apiKey    string
 	apiSecret string
+	// Timeout represents the default timeout for the async requests
+	Timeout time.Duration
+	// RetryStrategy represents the waiting strategy for polling the async requests
+	RetryStrategy RetryStrategyFunc
 }
+
+// RetryStrategyFunc represents a how much time to wait between two calls to CloudStack
+type RetryStrategyFunc func(int64) time.Duration
 
 // Topology represents a view of the servers
 type Topology struct {
