+++ virtual_machines_test.go
@@ -8,23 +8,23 @@ import (
 
 func TestVirtualMachines(t *testing.T) {
 	var _ Taggable = (*VirtualMachine)(nil)
-	var _ AsyncCommand = (*DeployVirtualMachine)(nil)
-	var _ AsyncCommand = (*DestroyVirtualMachine)(nil)
-	var _ AsyncCommand = (*RebootVirtualMachine)(nil)
-	var _ AsyncCommand = (*StartVirtualMachine)(nil)
-	var _ AsyncCommand = (*StopVirtualMachine)(nil)
-	var _ AsyncCommand = (*ResetPasswordForVirtualMachine)(nil)
-	var _ Command = (*UpdateVirtualMachine)(nil)
-	var _ Command = (*ListVirtualMachines)(nil)
-	var _ Command = (*GetVMPassword)(nil)
-	var _ AsyncCommand = (*RestoreVirtualMachine)(nil)
-	var _ Command = (*ChangeServiceForVirtualMachine)(nil)
-	var _ AsyncCommand = (*ScaleVirtualMachine)(nil)
-	var _ Command = (*RecoverVirtualMachine)(nil)
-	var _ AsyncCommand = (*ExpungeVirtualMachine)(nil)
-	var _ AsyncCommand = (*AddNicToVirtualMachine)(nil)
-	var _ AsyncCommand = (*RemoveNicFromVirtualMachine)(nil)
-	var _ AsyncCommand = (*UpdateDefaultNicForVirtualMachine)(nil)
+	var _ asyncCommand = (*DeployVirtualMachine)(nil)
+	var _ asyncCommand = (*DestroyVirtualMachine)(nil)
+	var _ asyncCommand = (*RebootVirtualMachine)(nil)
+	var _ asyncCommand = (*StartVirtualMachine)(nil)
+	var _ asyncCommand = (*StopVirtualMachine)(nil)
+	var _ asyncCommand = (*ResetPasswordForVirtualMachine)(nil)
+	var _ syncCommand = (*UpdateVirtualMachine)(nil)
+	var _ syncCommand = (*ListVirtualMachines)(nil)
+	var _ syncCommand = (*GetVMPassword)(nil)
+	var _ asyncCommand = (*RestoreVirtualMachine)(nil)
+	var _ syncCommand = (*ChangeServiceForVirtualMachine)(nil)
+	var _ asyncCommand = (*ScaleVirtualMachine)(nil)
+	var _ syncCommand = (*RecoverVirtualMachine)(nil)
+	var _ asyncCommand = (*ExpungeVirtualMachine)(nil)
+	var _ asyncCommand = (*AddNicToVirtualMachine)(nil)
+	var _ asyncCommand = (*RemoveNicFromVirtualMachine)(nil)
+	var _ asyncCommand = (*UpdateDefaultNicForVirtualMachine)(nil)
 }
 
 func TestVirtualMachine(t *testing.T) {
